Native App Phase 5 (5/8 follow-up): fix Windows clippy unnecessary_cast
CI 5.6 caught a latent Windows-only clippy lint that local macOS
clippy could not see: SW_NORMAL is already typed `i32` in
windows-sys, so the `as i32` cast in ShellExecuteW's nShowCmd arg
trips `clippy::unnecessary_cast`. On macOS the const doesn't exist
and the entire Windows block is cfg-gated out, so the cast never
compiled there - exactly the class of latent issue 5.6's
windows-latest CI lane was added to surface.

Drop the cast; behaviour identical, clippy green on all three OSes.

+++ app/src-tauri/src/elevation.rs
@@ -81,7 +81,8 @@ pub fn relaunch_self_as_admin() -> Result<(), String> {
             exe_w.as_ptr(),
             std::ptr::null(),
             std::ptr::null(),
-            SW_NORMAL as i32,
+            // `SW_NORMAL` is already `i32` in windows-sys; no cast.
+            SW_NORMAL,
         )
     };
     // Per Win32: a value > 32 means success; anything ≤ 32 is an
